Add files via upload
the attached PNG files is used by the script to recognise when the webpage has finished uploading the videos. They must be saved in the same folder as the python script. 
If this is not working for you, you may need to use SnagIt to capture the image, as the resolutions may be different from screen to screen